requirements.txt
click>=7.0
fastparquet>=0.3.3
boto3>=1.12.11,<1.17.45
flask>=1.1
genson==1.1.0
HiYaPyCo==0.4.16
PyAthena>=1.10
Pygments>=2.6.1
jsonschema>=2.5.1
requests>=2.9.1
pytest-spec==2.0.0
python-dotenv>=0.13.0
tabulate>=0.8.7
pandas>=1.0.2
s3fs>=0.4.2,<0.5.0
jsonlines>=1.2.0
fastparquet>=0.3.3
python-magic==0.4.18
dask-kubernetes==0.10.1
XlsxWriter>=1.3.1
setuptools>=40.3.0
lz4>=3.1.0
numpy==1.18.1
blosc==1.9.1
pyexcelerate==0.9.0
distributed==2.22.0
dask==2.22.0
numpy==1.18.1
pytest>=6.0.1
cloudpickle==1.5.0
connexion>=2.6
swagger-ui-bundle
connexion[swagger-ui]>2.6
mason_dask==0.1.26
Markdown==3.2.2
lambdas==0.1.0
returns==0.15.0
partd>=0.3.10
asciidag==0.1.0
file:///Users/kyle/dev/typistry#egg=typistry==0.0.7 --no-index

